+++ dsp/dsp-api/build.gradle.kts
@@ -14,6 +14,5 @@
  */
 
 dependencies {
-    api(project(":dsp:dsp-api"))
     implementation(project(":core"))
 }
