+++ dsp/dsp-api/src/main/java/org/eclipse/dataspacetck/dsp/system/api/connector/NegotiationListener.java
