+++ dsp/dsp-api/src/main/java/org/eclipse/dataspacetck/dsp/system/api/message/DspConstants.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.dataspacetck.dsp.system.message;
+package org.eclipse.dataspacetck.dsp.system.api.message;
 
 /**
  * DSP Message constants.
