+++ dsp/dsp-api/src/main/java/org/eclipse/dataspacetck/dsp/system/api/mock/ConsumerNegotiationMock.java
