+++ dsp/dsp-api/src/main/java/org/eclipse/dataspacetck/dsp/system/api/statemachine/LockManager.java
@@ -28,7 +28,7 @@ class LockManager {
 
     private final ReadWriteLock lock;
 
-    public LockManager() {
+    LockManager() {
         this.lock = new ReentrantReadWriteLock();
     }
