+++ dsp/dsp-contract-negotiation/build.gradle.kts
@@ -15,7 +15,8 @@
 
 dependencies {
     implementation(project(":core"))
-    implementation(project(":dsp:dsp-system"))
+    implementation(project(":dsp:dsp-api"))
+    testImplementation(project(":dsp:dsp-system"))
 }
 
 tasks.test {
