+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/DspSystemLauncher.java
@@ -21,16 +21,14 @@ import org.eclipse.dataspacetck.core.spi.system.ServiceConfiguration;
 import org.eclipse.dataspacetck.core.spi.system.ServiceResolver;
 import org.eclipse.dataspacetck.core.spi.system.SystemConfiguration;
 import org.eclipse.dataspacetck.core.spi.system.SystemLauncher;
-import org.eclipse.dataspacetck.dsp.system.api.client.ConsumerNegotiationClient;
-import org.eclipse.dataspacetck.dsp.system.api.client.ProviderNegotiationClient;
+import org.eclipse.dataspacetck.dsp.system.client.ConsumerNegotiationClient;
+import org.eclipse.dataspacetck.dsp.system.client.ProviderNegotiationClient;
 import org.eclipse.dataspacetck.dsp.system.api.connector.Connector;
 import org.eclipse.dataspacetck.dsp.system.api.connector.Consumer;
 import org.eclipse.dataspacetck.dsp.system.api.mock.ConsumerNegotiationMock;
 import org.eclipse.dataspacetck.dsp.system.api.mock.ProviderNegotiationMock;
 import org.eclipse.dataspacetck.dsp.system.api.pipeline.ConsumerNegotiationPipeline;
-import org.eclipse.dataspacetck.dsp.system.pipeline.ConsumerNegotiationPipelineImpl;
 import org.eclipse.dataspacetck.dsp.system.api.pipeline.ProviderNegotiationPipeline;
-import org.eclipse.dataspacetck.dsp.system.pipeline.ProviderNegotiationPipelineImpl;
 import org.eclipse.dataspacetck.dsp.system.client.ConsumerNegotiationClientImpl;
 import org.eclipse.dataspacetck.dsp.system.client.ProviderNegotiationClientImpl;
 import org.eclipse.dataspacetck.dsp.system.connector.TckConnector;
@@ -38,6 +36,8 @@ import org.eclipse.dataspacetck.dsp.system.mock.ConsumerNegotiationMockImpl;
 import org.eclipse.dataspacetck.dsp.system.mock.NoOpConsumerNegotiationMock;
 import org.eclipse.dataspacetck.dsp.system.mock.NoOpProviderNegotiationMock;
 import org.eclipse.dataspacetck.dsp.system.mock.ProviderNegotiationMockImpl;
+import org.eclipse.dataspacetck.dsp.system.pipeline.ConsumerNegotiationPipelineImpl;
+import org.eclipse.dataspacetck.dsp.system.pipeline.ProviderNegotiationPipelineImpl;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
