+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/client/ProviderNegotiationClient.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.dataspacetck.dsp.system.api.client;
+package org.eclipse.dataspacetck.dsp.system.client;
 
 import java.util.Map;
